+++ config/routes.js
@@ -3,6 +3,7 @@ Router.add({
 	methods    : 'post',
 	paths      : '/api/form/data/related',
 	policy     : 'logged_in',
+	is_system_route : true,
 });
 
 Router.add({
@@ -10,4 +11,14 @@ Router.add({
 	methods    : 'post',
 	paths      : '/api/form/data/qbdata',
 	policy     : 'logged_in',
+	is_system_route : true,
+});
+
+Router.add({
+	name            : 'FormApi#recompute',
+	methods         : 'post',
+	paths           : '/api/form/data/recompute/{model_name}/{field}',
+	policy          : 'logged_in',
+	is_system_route : true,
+	permission      : 'model.{model_name}.recompute.{field}',
 });