+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Fix `al-toggle` not firing a change event
 * Fixed `al-field` rendering to update content for non-value setter elements when a new value is assigned
 * Use `constraints` data from sibling fields when loading associated field data
+* Automatically update computed field values when a field it depends on changes
 
 ## 0.2.8 (2023-10-15)